refactor: Update Next to 15, disable (temp) maps
