+++ app/home/[id]/_components/HomeOnMap.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Skeleton } from "@/components/ui/skeleton";
 import dynamic from "next/dynamic";
 import { FC } from "react";
