+++ app/layout.tsx
@@ -1,19 +1,32 @@
 import type { Metadata } from "next";
-import { PropsWithChildren } from "react";
-import { Inter } from "next/font/google";
+import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { Navbar } from "./_components";
 
-const inter = Inter({ subsets: ["latin"] });
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+  preload: false,
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+  preload: false,
+});
 
 export const metadata: Metadata = {
   title: "Trip25 || BNB, Photos, Maps",
   description: "Next 14 + Prisma + Tailwind + ShadcnUI Trips App",
 };
 
-const RootLayout = ({ children }: PropsWithChildren) => (
-  <html lang="en">
-    <body className={inter.className}>
+const RootLayout = ({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) => (
+  <html lang="en" suppressHydrationWarning>
+    <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
       <Navbar />
       {children}
     </body>
