+++ package.json
@@ -30,7 +30,7 @@
     "leaflet": "^1.9.4",
     "lucide-react": "^0.378.0",
     "mapbox-gl": "2.14.1",
-    "next": "^14.2.24",
+    "next": "^15.2.2",
     "react": "^18.2.0",
     "react-date-range": "^2.0.1",
     "react-dom": "^18.2.0",
